Renamed modules and removed stuff that is headed to a private repo/package.
