+++ ng-package.json
@@ -1,7 +1,7 @@
 {
   "$schema": "./node_modules/ng-packagr/package.schema.json",
   "name": "wily-angular-commons",
-  "version": "7.0.8",
+  "version": "7.0.15",
   "ngPackage": {
     "lib": {
       "entryFile": "public_api.ts"
