--- src/app/modules/help-widget/help-widget.component.html
@@ -1,46 +1,0 @@
-<wily-dialog [title]="help ? help.name : ''" [object]="help" #dialog>
-
-  <div style="height:97%;" *ngIf="help">
-    <!-- LOADER -->
-    <div *ngIf="status === EndpointStatus.LOADING" class="center_center" style="text-align:center;height: 100%;">
-      <div>
-        <img src="https://styles.transcend.csra.io/v2.0.2/assets/loading.gif" alt="Loading..." />
-      </div>
-    </div>
-    <!-- END LOADER -->
-
-    <!-- HELP CONTENT -->
-    <div *ngIf="status === EndpointStatus.SUCCESS">
-      <div class="row space_between v_center">
-        <div class="wac_help_description overlay" *ngIf="help.description" style="border:unset">
-          {{help.description}}
-        </div>
-      </div>
-      <div class="ql-editor wac_help_text" [innerHTML]="help.body"></div>
-    </div>
-    <!-- END HELP CONTENT -->
-
-    <!-- UNAVAILABLE -->
-    <div *ngIf="status === EndpointStatus.ERROR" class="center_center" style="text-align:center;height: 100%;">
-      <div>
-        <div class="wac_help_notavail">This help topic is currently unavailable, please check back later.</div>
-        <div class="center_center" *ngIf="canEdit">
-          <div class="overlay" style="margin-top:15px;padding:5px;">
-            <div>Create this topic by visiting Application Management and creating a new help entry</div>
-            <div>with the key, <strong>{{helpKey}}</strong> in the product <strong>{{productKey}}</strong></div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <!-- END UNAVAILABLE -->
-
-  </div>
-
-</wily-dialog>
-
-<div class="action" (click)="showHelp()" role="button" aria-label="Show Help">
-  <ng-content *ngIf="displayMode !== 'default'"></ng-content>
-  <div class="btn_pulse" [ngClass]="sizeClass" *ngIf="displayMode === 'default'">
-    <i class="fas fa-question fa-fw"></i>
-  </div>
-</div>
