--- src/app/modules/help-widget/help-widget.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { HelpWidgetComponent } from './help-widget.component';
-
-describe('HelpWidgetComponent', () => {
-  let component: HelpWidgetComponent;
-  let fixture: ComponentFixture<HelpWidgetComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ HelpWidgetComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(HelpWidgetComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
