+++ src/app/modules/icons/wily-icons-light.model.ts
