+++ src/app/modules/icons/wily-icons-regular.model.ts
