+++ src/app/modules/icons/wily-icons-solid.model.ts
