+++ src/app/modules/icons/wily-icons.module.ts
