+++ src/app/modules/safe-pipes/wily-safe-pipes.module.ts
@@ -16,5 +16,5 @@ import {SafeHtmlPipe} from './safe-html.pipe';
     SafeStylePipe
   ]
 })
-export class SafePipesModule {
+export class WilySafePipesModule {
 }
