--- src/app/modules/theming/components/theming.component.html
@@ -1,5 +1,0 @@
-<select aria-label="Change Theme" [ngClass]="selectClass" [(ngModel)]="theme" (change)="changeTheme()">
-  <option *ngFor="let optionTheme of themes" [ngValue]="optionTheme" [selected]="optionTheme.themeId === theme.themeId">
-    {{optionTheme.name}}{{optionTheme.isDefault === 'Y' ? ' - Default' : ''}}
-  </option>
-</select>
