--- src/app/modules/theming/theming.module.spec.ts
@@ -1,13 +1,0 @@
-import { ThemingModule } from './theming.module';
-
-describe('ThemingModule', () => {
-  let themingModule: ThemingModule;
-
-  beforeEach(() => {
-    themingModule = new ThemingModule();
-  });
-
-  it('should create an instance', () => {
-    expect(themingModule).toBeTruthy();
-  });
-});
