+++ src/public_api.ts
@@ -1,25 +1,21 @@
 /** MODULES */
-export * from './app/modules/background-carousel/background-carousel.module';
-export * from './app/modules/help-widget/help-widget.module';
-export * from './app/modules/icon-select/icon-select.module';
-export * from './app/modules/paginator/paginator.module';
-export * from './app/modules/profile-pic/profile-pic.module';
-export * from './app/modules/push-container/push-container.module';
-export * from './app/modules/rich-text/rich-text.module';
-export * from './app/modules/safe-pipes/safe-pipes.module';
-export * from './app/modules/theming/theming.module';
-export * from './app/modules/wily-dialog/wily-dialog.module';
-export * from './app/modules/wily-icons/wily-icons.module';
+export * from './app/modules/background-carousel/wily-background-carousel.module';
+export * from './app/modules/icon-select/wily-icon-select.module';
+export * from './app/modules/paginator/wily-paginator.module';
+export * from './app/modules/profile-pic/wily-profile-pic.module';
+export * from './app/modules/push-container/wily-push-container.module';
+export * from './app/modules/rich-text/wily-rich-text.module';
+export * from './app/modules/safe-pipes/wily-safe-pipes.module';
+export * from './app/modules/dialog/wily-dialog.module';
+export * from './app/modules/icons/wily-icons.module';
 
 /** COMPONENTS */
-export * from './app/modules/help-widget/help-widget.component';
 export * from './app/modules/icon-select/icon-select.component';
 export * from './app/modules/paginator/paginator.component';
 export * from './app/modules/profile-pic/profile-pic.component';
 export * from './app/modules/push-container/push-container.component';
 export * from './app/modules/rich-text/rich-text.component';
-export * from './app/modules/theming/components/theming.component';
-export * from './app/modules/wily-dialog/dialog.component';
+export * from './app/modules/dialog/dialog.component';
 
 /** INTERCEPTORS */
 export * from './app/shared/interceptors/json.interceptor';
@@ -29,11 +25,9 @@ export * from './app/shared/interceptors/cognito-auth.interceptor';
 export * from './app/shared/services/base-data.service';
 export * from './app/shared/services/local-storage.service';
 export * from './app/shared/services/cognito-auth.service';
-export * from './app/modules/theming/services/theming.service';
 
 /** MODELS */
 export * from './app/modules/paginator/pagination-event.model';
-export * from './app/modules/theming/models/theme.model';
 export * from './app/shared/models/auditable.model';
 export * from './app/shared/models/data-list.model';
 export * from './app/shared/models/id.model';
